+++ shared/chat/index.stories.js
@@ -4,7 +4,7 @@ import createChannel from './create-channel/index.stories'
 import inboxRow from './inbox/row/index.stories'
 import infoPanel from './conversation/info-panel/index.stories'
 import infoPanelNotifications from './conversation/info-panel/notifications/index.stories'
-import inputArea from './conversation/input-area/index.stories'
+import inputArea from './conversation/input-area/normal/index.stories'
 import manageChannels from './manage-channels/index.stories'
 import sendAnimation from './conversation/messages/wrapper/chat-send.stories'
 import setExplodingMessage from './conversation/messages/set-explode-popup/index.stories'
